requirements_python36_ubuntu18.txt
absl-py==1.0.0
astor==0.8.1
cached-property==1.5.2
dataclasses==0.8
gast==0.2.2
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.1.0
importlib-metadata==4.8.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.6
numpy==1.19.5
opencv-python==4.5.5.64
opt-einsum==3.3.0
pkg_resources==0.0.0
protobuf==3.19.4
six==1.16.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
tf-slim==1.1.0
typing_extensions==4.1.1
Werkzeug==2.0.3
wrapt==1.14.0
zipp==3.6.0
